A bit of UFS spring cleaning - move all UFS source into one folder
