+++ sys/src/9/ufs/dirhash.h
