+++ sys/src/9/ufs/ffs_balloc.c
@@ -60,7 +60,7 @@
  */
 
 #include "u.h"
-#include "../../port/lib.h"
+#include "../port/lib.h"
 #include "mem.h"
 #include "dat.h"
 #include "fns.h"
