+++ sys/src/9/ufs/ffs_extern.h
