+++ sys/src/9/ufs/ffs_inode.c
@@ -30,7 +30,7 @@
  */
 
 #include "u.h"
-#include "../../port/lib.h"
+#include "../port/lib.h"
 #include "mem.h"
 #include "dat.h"
 #include "fns.h"
