+++ sys/src/9/ufs/ffs_rawread.c
