+++ sys/src/9/ufs/ffs_suspend.c
