+++ sys/src/9/ufs/gjournal.h
