+++ sys/src/9/ufs/softdep.h
