+++ sys/src/9/ufs/ufs.json
@@ -4,35 +4,31 @@
 			"-fasm",
 			"-I",
 			"../ufs",
-			"-I",
-			"../ufs/ffs",
-			"-I",
-			"../ufs/ufs",
 			"-DNO_FFS_SNAPSHOT"
 		],
 		"#SourceFiles": [
-			"../ufs/ffs/ffs_rawread.c",
-			"../ufs/ufs/ufs_acl.c",
-			"../ufs/ufs/ufs_bmap.c",
-			"../ufs/ufs/ufs_dirhash.c",
-			"../ufs/ufs/ufs_extattr.c",
-			"../ufs/ufs/ufs_gjournal.c",
-			"../ufs/ufs/ufs_quota.c"
+			"../ufs/ffs_rawread.c",
+			"../ufs/ufs_acl.c",
+			"../ufs/ufs_bmap.c",
+			"../ufs/ufs_dirhash.c",
+			"../ufs/ufs_extattr.c",
+			"../ufs/ufs_gjournal.c",
+			"../ufs/ufs_quota.c"
 		],
 		"SourceFiles": [
-			"../ufs/ffs/ffs_alloc.c",
-			"../ufs/ffs/ffs_balloc.c",
-			"../ufs/ffs/ffs_inode.c",
-			"../ufs/ffs/ffs_snapshot.c",
-			"../ufs/ffs/ffs_softdep.c",
-			"../ufs/ffs/ffs_subr.c",
-			"../ufs/ffs/ffs_suspend.c",
-			"../ufs/ffs/ffs_vfsops.c",
-			"../ufs/ffs/ffs_vnops.c",
-			"../ufs/ufs/ufs_inode.c",
-			"../ufs/ufs/ufs_lookup.c",
-			"../ufs/ufs/ufs_vfsops.c",
-			"../ufs/ufs/ufs_vnops.c",
+			"../ufs/ffs_alloc.c",
+			"../ufs/ffs_balloc.c",
+			"../ufs/ffs_inode.c",
+			"../ufs/ffs_snapshot.c",
+			"../ufs/ffs_softdep.c",
+			"../ufs/ffs_subr.c",
+			"../ufs/ffs_suspend.c",
+			"../ufs/ffs_vfsops.c",
+			"../ufs/ffs_vnops.c",
+			"../ufs/ufs_inode.c",
+			"../ufs/ufs_lookup.c",
+			"../ufs/ufs_vfsops.c",
+			"../ufs/ufs_vnops.c",
 			"../ufs/ufs_harvey.c",
 			"../ufs/ufs_mountpoint.c"
 		]
