+++ sys/src/9/ufs/ufs_bmap.c
