+++ sys/src/9/ufs/ufs_dirhash.c
