+++ sys/src/9/ufs/ufs_extern.h
