+++ sys/src/9/ufs/ufs_gjournal.c
