+++ sys/src/9/ufs/ufs_mountpoint.c
@@ -10,7 +10,7 @@
 
 
 #include "u.h"
-#include "../../port/lib.h"
+#include "../port/lib.h"
 #include "mem.h"
 #include "dat.h"
 #include "fns.h"
